\name{DataExp}
\alias{DataExp}
\docType{data}
\title{DataExp}
\description{Example of data collected via servey reserch}

\usage{data(DataExp)}
\format{
  A data frame with 112 observations on the following 39 variables.
  \describe{
    \item{\code{ID}}{a factor with levels \code{Agencje} \code{Firmy} \code{Mieszkancy} \code{Realizatorzy} \code{Uczelnie} \code{Urzedy}}
    \item{\code{pyt1}}{a numeric vector}
    \item{\code{pyt2}}{a numeric vector}
    \item{\code{pyt3}}{a numeric vector}
    \item{\code{pyt4}}{a numeric vector}
    \item{\code{pyt5}}{a numeric vector}
    \item{\code{pyt6}}{a numeric vector}
    \item{\code{pyt7}}{a numeric vector}
    \item{\code{pyt8}}{a numeric vector}
    \item{\code{pyt9}}{a numeric vector}
    \item{\code{pyt10}}{a numeric vector}
    \item{\code{pyt11}}{a numeric vector}
    \item{\code{pyt12}}{a numeric vector}
    \item{\code{pyt13}}{a numeric vector}
    \item{\code{pyt14}}{a numeric vector}
    \item{\code{pyt15}}{a numeric vector}
    \item{\code{pyt16}}{a numeric vector}
    \item{\code{prof1}}{a numeric vector}
    \item{\code{prof2}}{a numeric vector}
    \item{\code{prof3}}{a numeric vector}
    \item{\code{prof4}}{a numeric vector}
    \item{\code{prof5}}{a numeric vector}
    \item{\code{cost1}}{a numeric vector}
    \item{\code{cost2}}{a numeric vector}
    \item{\code{cost3}}{a numeric vector}
    \item{\code{cost4}}{a numeric vector}
    \item{\code{cost5}}{a numeric vector}
    \item{\code{my1}}{a numeric vector}
    \item{\code{my2}}{a numeric vector}
    \item{\code{my3}}{a numeric vector}
    \item{\code{my4}}{a numeric vector}
    \item{\code{my5}}{a numeric vector}
    \item{\code{my6}}{a numeric vector}
    \item{\code{me1}}{a numeric vector}
    \item{\code{me2}}{a numeric vector}
    \item{\code{me3}}{a numeric vector}
    \item{\code{me4}}{a numeric vector}
    \item{\code{me5}}{a numeric vector}
    \item{\code{me6}}{a numeric vector}
  }
}
\examples{data(DataExp)}
\keyword{datasets}
