\name{StakeholderAnalysis-package}
\alias{StakeholderAnalysis-package}
\alias{StakeholderAnalysis}
\docType{package}
\title{
\packageTitle{StakeholderAnalysis}
}
\description{Proposes an original instrument for measuring stakeholder influence on the development of an infrastructure project that is carried through by a municipality.
\packageDescription{StakeholderAnalysis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{StakeholderAnalysis}
\packageIndices{StakeholderAnalysis}
}
\author{Lech Kujawski
\packageAuthor{StakeholderAnalysis}

Maintainer: \packageMaintainer{StakeholderAnalysis}
}
\references{
Hester, P.T., & Adams, K.M. (2013) <doi:10.1016/j.procs.2013.09.282>
Hester, P.T., Bradley, J.M., MacGregor K.A. (2012) <doi:10.1504/IJSSE.2012.052687>
}
\keyword{packages}
