\name{StakeholderAnalysis-package}
\alias{StakeholderAnalysis-package}
\alias{StakeholderAnalysis}
\docType{package}
\title{
\packageTitle{StakeholderAnalysis}
}
\description{This project was financed by The National Center of Reaserch and Development (gran number IS-2/88/NCBR/2015). This software is an original instrument for measuring stakeholder influence on the development of a publicly-funded infrastructure project and for estimating its cost and benefits, as preceived by different stakeholder groups. Its originality lies in applicatin of Leontief's input-output analysis to estimating both stakeholder influence on a infrastructure project and its cost and benefits, as perceived by different stakeholder groups. Admittedly, in the literature there are studies that draw on Leontief's modle to estimate stakeholder influence or, separately, to measure a project's percived costs and benefits. That said, none of this reserch work - unlike our package - combines the two focuses. It follows that our software, uniquely, links together Leontief's input-output analysis, stakeholder influence measurement and estimation of a project's costs and benefits.
Therefore, it constitutes a useful instrument that might be of particular interest to managers and municipality official responsible for implementation of a large-scale infrastructure projects.
\packageDescription{StakeholderAnalysis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{StakeholderAnalysis}
\packageIndices{StakeholderAnalysis}
}
\author{Lech Kujawski
\packageAuthor{StakeholderAnalysis}

Maintainer: \packageMaintainer{StakeholderAnalysis}
}
\references{
Hester, P.T., & Adams, K.M. (2013) <doi:10.1016/j.procs.2013.09.282>
Hester, P.T., Bradley, J.M., MacGregor K.A. (2012) <doi:10.1504/IJSSE.2012.052687>
}
\keyword{packages}
