\name{DataExp}
\alias{DataExp}
\docType{data}
\title{DataExp}
\description{Example of data collected via servey reserch}

\usage{data(DataExp)}
\format{
  A data frame with 112 observations on the following 39 variables.
  \describe{
    \item{\code{ID}}{a factor with levels \code{Consulting Agencies} \code{Firms} \code{Local Residents} \code{Project Implementers} \code{Research Units} \code{Local Authorities}}
    \item{\code{pyt1}}{a numeric vector on a five-point Likert scale. I have a positive view on the implementation of the project (Attitude).}
    \item{\code{pyt2}}{a numeric vector on a five-point Likert scale. I have effective power over the implementation of the project (Power).}
    \item{\code{pyt3}}{a numeric vector on a five-point Likert scale. I think that the implementation of the project is urgent (Urgency).}
    \item{\code{pyt4}}{a numeric vector on a five-point Likert scale. I am entitled to be consulted about the implementation of the project (Legitimacy).}
    \item{\code{pyt5}}{a numeric vector on a five-point Likert scale. The project should be implemented as soon as possible (Urgency).}
    \item{\code{pyt6}}{a numeric vector on a five-point Likert scale. It is justified to consult me about the implementation of the project (Legitimacy).}
    \item{\code{pyt7}}{a numeric vector on a five-point Likert scale. Changes to the implementation of the project depends on me (Power).}
    \item{\code{pyt8}}{a numeric vector on a five-point Likert scale. Given my role in the project, I should bo consulted abouts its implementation (Legitimacy).}
    \item{\code{pyt9}}{a numeric vector on a five-point Likert scale. In my opinion, the implementaton of the project is a matter of urgency (Urgency).}
    \item{\code{pyt10}}{a numeric vector on a five-point Likert scale. My attitude to the implementation of the project is unambiguously positive (Attitude).}
    \item{\code{pyt11}}{a numeric vector on a five-point Likert scale. The implementation of the project is very important to me (Urgency).}
    \item{\code{pyt12}}{a numeric vector on a five-point Likert scale. Implementers of the project can count on me for support (Attitude).}
    \item{\code{pyt13}}{a numeric vector on a five-point Likert scale. My power over the implantation of the project is considerable (Power).}
    \item{\code{pyt14}}{a numeric vector on a five-point Likert scale. I support the implantation of the project (Attitude).}
    \item{\code{pyt15}}{a numeric vector on a five-point Likert scale. I am in a position to influence the implantation of the project (Power).}
    \item{\code{pyt16}}{a numeric vector on a five-point Likert scale. My role in the project justifies consulting me about its implementation (Legitimacy).}
    \item{\code{prof1}}{a numeric vector on a five-point Likert scale. The implementation of the project will generate new jobs.}
    \item{\code{prof2}}{a numeric vector on a five-point Likert scale. The implementation of the project will enhance the technological potential of the region.}
    \item{\code{prof3}}{a numeric vector on a five-point Likert scale. Thanks to the implementation of the project, more results of scientific research will be put into practice.}
    \item{\code{prof4}}{a numeric vector on a five-point Likert scale. The implementation of the project will translate into a greater number of patents and inventions.}
    \item{\code{prof5}}{a numeric vector on a five-point Likert scale. Thanks to the implementation of the project, local infrastructure will be modernized.}
    \item{\code{cost1}}{a numeric vector on a five-point Likert scale. The implementation of the project will increase road traffic in the area.}
    \item{\code{cost2}}{a numeric vector on a five-point Likert scale. The implementation of the project will negatively affect the environment.}
    \item{\code{cost3}}{a numeric vector on a five-point Likert scale. The implementation of the project will increase air pollution.}
    \item{\code{cost4}}{a numeric vector on a five-point Likert scale. The implementation of the project will increase noise.}
    \item{\code{cost5}}{a numeric vector on a five-point Likert scale. The implementation of the project will make it harder to park a car.}
    \item{\code{my1}}{a numeric vector on a five-point Likert scale. I have influence over Consulting Agencies.}
    \item{\code{my2}}{a numeric vector on a five-point Likert scale. I have influence over Firms.}
    \item{\code{my3}}{a numeric vector on a five-point Likert scale. I have influence over Local Residents.}
    \item{\code{my4}}{a numeric vector on a five-point Likert scale. I have influence over Project Implementers.}
    \item{\code{my5}}{a numeric vector on a five-point Likert scale. I have influence over Research Units.}
    \item{\code{my6}}{a numeric vector on a five-point Likert scale. I have influence over Local Authorities.}
    \item{\code{me1}}{a numeric vector on a five-point Likert scale. Consulting Agencies have influence over me.}
    \item{\code{me2}}{a numeric vector on a five-point Likert scale. Firms have influence over me.}
    \item{\code{me3}}{a numeric vector on a five-point Likert scale. Local Residents have influence over me.}
    \item{\code{me4}}{a numeric vector on a five-point Likert scale. Project Implementers have influence over me.}
    \item{\code{me5}}{a numeric vector on a five-point Likert scale. Research Units have influence over me.}
    \item{\code{me6}}{a numeric vector on a five-point Likert scale. Local Authorities have influence over me.}
  }
}
\examples{data(DataExp)}
\keyword{datasets}
